// src/lib/constants/enums/skill-names.ts
export enum SkillNames {
    AGILITY = 'agility',
    ATTACK = 'attack',
    CONSTRUCTION = 'construction',
    COOKING = 'cooking',
    CRAFTING = 'crafting',
    DEFENCE = 'defence',
    FARMING = 'farming',
    FIREMAKING = 'firemaking',
    FISHING = 'fishing',
    FLETCHING = 'fletching',
    HERBLORE = 'herblore',
    HITPOINTS = 'hitpoints',
    HUNTER = 'hunter',
    MAGIC = 'magic',
    MINING = 'mining',
    PRAYER = 'prayer',
    RANGED = 'ranged',
    RUNECRAFTING = 'runecrafting',
    SLAYER = 'slayer',
    SMITHING = 'smithing',
    STRENGTH = 'strength',
    THIEVING = 'thieving',
    WOODCUTTING = 'woodcutting',
}